+++ NEWS.md
@@ -1,3 +1,11 @@
+# mixedup 0.1.1
+
+* Add glmmTMB method for for extract_vc
+
+# mixedup 0.1.1
+
+* Add initial functionality for extracting variance components
+
 # mixedup 0.1.0
 
 * A couple functions for a few packages